fix(kg-editor): keep route exports Next-compatible
